% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{classify}
\alias{classify}
\title{classify}
\usage{
classify(d, id = NULL, vars = NULL, ...)
}
\arguments{
\item{d}{data.frame}

\item{id}{column id}

\item{vars}{variables}

\item{...}{passed on to melt}
}
\value{
data.frame
}
\description{
relabel factors
}
\details{
Wide to long format data.frame with new factor variable(s) describing the original columns
}
\seealso{
Other helping_functions: 
\code{\link{internal_field}()},
\code{\link{invert_stack}()},
\code{\link{lfief}()},
\code{\link{modify_levels}()}
}
\author{
Baptiste Auguie
}
\concept{helping_functions}
