% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic.r
\name{combine_layer}
\alias{combine_layer}
\title{combine_layer}
\usage{
combine_layer(r1, r2, kd)
}
\arguments{
\item{r1}{reflection coefficient}

\item{r2}{reflection coefficient}

\item{kd}{k*d}
}
\value{
combined complex reflectivity
}
\description{
combine layer
}
\details{
reflection coefficient for a layer
}
\author{
baptiste Auguie
}
\concept{user_level}
