% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic.r
\name{dbr_analytic}
\alias{dbr_analytic}
\title{dbr_analytic}
\usage{
dbr_analytic(wavelength, lambda0, n1, n2, nleft, d1 = lambda0/4/n1,
  d2 = lambda0/4/n2, ...)
}
\arguments{
\item{wavelength}{in nm}

\item{lambda0}{central wavelength of the stopband}

\item{n1}{real refractive index for odd layers}

\item{n2}{real refractive index for even layers}

\item{nleft}{real refractive index for incident medium}

\item{d1}{odd layer thickness in nm}

\item{d2}{even layer thickness in nm}

\item{...}{ignored}
}
\value{
data.frame with complex reflectivity
}
\description{
semi-infinite DBR
}
\details{
periodic structure of dielectric layers
}
\note{
issue at lambda0/2 needs investigating
}
\author{
baptiste Auguie
}
\references{
Amir and Vukusic, 2013, arXiv:1209.3776v2
}

