% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.r
\name{embed_stack}
\alias{embed_stack}
\title{embed_stack}
\usage{
embed_stack(s, nleft = 1, nright = 1, dleft = 200, dright = 200, ...)
}
\arguments{
\item{s}{stack (finite structure)}

\item{nleft}{real refractive index on the left side}

\item{nright}{real refractive index on the right side}

\item{dleft}{dummy layer thickness in nm}

\item{dright}{dummy layer thickness in nm}

\item{...}{ignored}
}
\value{
list of class 'stack'
}
\description{
Embed stack structure
}
\details{
embeds a stack in semi-infinite media
}
\seealso{
Other stack user_level: 
\code{\link{dbr_stack}()},
\code{\link{layer_stack}()},
\code{\link{tamm_stack_ir}()},
\code{\link{tamm_stack_porous}()},
\code{\link{tamm_stack}()}
}
\author{
baptiste Auguie
}
\concept{stack user_level}
