\name{epsilon_dispersion}
\alias{epsilon_dispersion}
\title{epsilon_dispersion}
\usage{
epsilon_dispersion(epsilon, wavelength = seq(400, 1000))
}
\arguments{
  \item{epsilon}{list of real or complex values}

  \item{wavelength}{numeric vector}
}
\value{
list
}
\description{
epsilon_dispersion
}
\details{
apply a function to a range of wavelength and return
dielectric function
}
\author{
baptiste Auguie
}

