% Generated by roxygen2 (4.0.0): do not edit by hand

\name{epsilon_label}
\alias{epsilon_label}
\title{epsilon_label}
\usage{
epsilon_label(epsilon = list(3.5, 1, 3, 1, "epsAu", 3, 3.5), names = NULL)
}
\arguments{
  \item{epsilon}{list of real or complex values}

  \item{names}{optional unique character names in order of
  appearance}
}
\value{
factor
}
\description{
epsilon_label
}
\details{
characterise the layers of a structure with unique labels
for metals and dielectrics
}
\author{
baptiste Auguie
}

