\name{epsilon_medium}
\alias{epsilon_medium}
\title{epsilon_medium}
\usage{
epsilon_medium(epsilon = list(3.5, 1, 3, 1, 12 + (0+1i), 3, 3.5))
}
\arguments{
  \item{epsilon}{}
}
\value{
factor
}
\description{
epsilon_medium
}
\details{
characterise the layers of a structure
}
\author{
baptiste Auguie
}

