% Generated by roxygen2 (4.0.0): do not edit by hand

\docType{data}
\name{gaussian}
\alias{gaussian}
\title{Rcpp module: gaussian}
\description{
Exposes C++ function integrand_gb
}
\details{
\itemize{ \item{integrand_gb}{ integrand for gaussian beam
excitation at a planar interface} }
}
\examples{
show( gaussian )
}

