\name{integrand1GL}
\alias{integrand1GL}
\title{integrand1GL}
\usage{integrand1GL(d=10, u, lambda, epsilon=list(incident = 1.5^2, 1^2),
    thickness=c(0, 0))}
\description{Dipole decay rates near a multilayer interface}
\details{Integrand of the dipole decay rates near a multilayer interface}
\author{baptiste Auguie}
\arguments{\item{d}{distance in nm}
\item{u}{transformed normalised in-plane wavevector sqrt(1-q^2)}
\item{lambda}{wavelength in nm}
\item{epsilon}{list of dielectric functions}
\item{thickness}{list of layer thicknesses}}
