\name{invert_incidence}
\alias{invert_incidence}
\title{invert_incidence}
\usage{
  invert_incidence(p)
}
\arguments{
  \item{p}{list}
}
\value{
  list
}
\description{
  invert the description of a multilayer to simulate the
  opposite direction of incidence
}
\details{
  inverts list of epsilon and thickness of layers
}
\author{
  Baptiste Auguie
}
\seealso{
  Other helping_functions: \code{\link{field.profile}}
}

