% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.r
\name{invert_stack}
\alias{invert_stack}
\title{invert_stack}
\usage{
invert_stack(p)
}
\arguments{
\item{p}{list}
}
\value{
list
}
\description{
invert the description of a multilayer to simulate the opposite direction of incidence
}
\details{
inverts list of epsilon and thickness of layers
}
\author{
Baptiste Auguie
}
\seealso{
Other helping_functions: \code{\link{classify}},
  \code{\link{internal_field}}, \code{\link{lfief}},
  \code{\link{modify_levels}}
}

