% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacks.r
\name{layer_stack}
\alias{layer_stack}
\title{layer_stack}
\usage{
layer_stack(epsilon = "epsAu", thickness = 50, ...)
}
\arguments{
\item{epsilon}{dielectric function (numeric, character, or complex)}

\item{thickness}{layer thickness in nm}

\item{...}{ignored}
}
\value{
list of class 'stack'
}
\description{
Single-layer stack structure
}
\details{
returns a stack describing a single layer
}
\author{
baptiste Auguie
}
\seealso{
Other stack user_level: \code{\link{dbr_stack}};
  \code{\link{embed_stack}}; \code{\link{tamm_stack_ir}};
  \code{\link{tamm_stack_porous}}; \code{\link{tamm_stack}}
}

