% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{modify_levels}
\alias{modify_levels}
\title{modify_levels}
\usage{
modify_levels(f, modify = list())
}
\arguments{
\item{f}{factor}

\item{modify}{named list}
}
\value{
factor
}
\description{
relabel factors
}
\author{
Baptiste Auguie
}
\seealso{
Other helping_functions: \code{\link{classify}};
  \code{\link{field_profile}}, \code{\link{lfief}};
  \code{\link{internal_field}}; \code{\link{invert_stack}}
}

