% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/multilayer.r
\name{cpp_multilayer}
\alias{cpp_multilayer}
\alias{cpp_multilayer_field}
\alias{multilayer}
\title{Transfer matrix formalism for multilayer stacks}
\usage{
cpp_multilayer(k0, kx, epsilon, thickness, z, psi, intensity)

cpp_multilayer_field(k0, kx, epsilon, thickness, z, psi)

multilayer(wavelength = 2 * pi/k0, k0 = 2 * pi/wavelength,
  angle = asin(q), q = sin(angle), epsilon = list(incident = 1.5^2, 1.33),
  thickness = c(0, 0), polarisation = c("p", "s"), d = 1, dout = d, ...)
}
\arguments{
\item{k0}{vector of wavenumbers}

\item{kx}{complex matrix of incident k parallel}

\item{epsilon}{complex matrix of epsilon}

\item{thickness}{vector of thicknesses}

\item{z}{vector of positions}

\item{psi}{scalar polarisation angle}

\item{intensity}{boolean: also return intensities}

\item{wavelength}{[vector] wavelength in nm}

\item{angle}{[vector] incident angles in radians}

\item{q}{[vector] normalised incident in-plane wavevector}

\item{polarisation}{[character] switch between p- and s- polarisation}

\item{d}{vector of distances where LFIEF are evaluated from each interface}

\item{dout}{vector of distances where LFIEF are evaluated outside the stack}

\item{...}{unused}

\item{k0}{vector of wavenumbers}

\item{kx}{complex matrix of incident k parallel}

\item{epsilon}{complex matrix of epsilon}

\item{thickness}{vector of thicknesses}

\item{z}{vector of positions}

\item{psi}{scalar polarisation angle}

\item{k0}{[vector] wavevector in nm^-1}

\item{epsilon}{list of N+2 dielectric functions, each of length 1 or length(wavelength)}

\item{thickness}{vector of N+2 layer thicknesses, first and last are dummy}
}
\value{
list with rs, rp, ts, tp, I (optional)

list with rs, rp, ts, tp, I

fresnel coefficients and field profiles
}
\description{
Full calculation for multiple wavelengths

Field calculation for one wavelength

Multilayer Fresnel coefficients
}
\details{
solves the EM problem of a multilayered interface
}
\section{Functions}{
\itemize{
\item \code{cpp_multilayer}: full multi

\item \code{cpp_multilayer_field}: full multi
}}
\author{
baptiste Auguie
}
\references{
Principles of surface-enhanced Raman spectroscopy and related plasmonic effects.
Eric C. Le Ru and Pablo G. Etchegoin, published by Elsevier, Amsterdam (2009).
}

