% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{palettes}
\alias{palettes}
\title{Colour palettes for multilayer stacks}
\format{Three palettes (vectors of colours)
\describe{
  \item{palette_dbr}{useful for a DBR}
  \item{palette_tamm}{useful for a DBR with Au coating}
  \item{palette_tamm_porous}{alternative Tamm colour scheme}
}}
\source{
See RColorBrewer package
}
\description{
Custom palettes for DBR and Tamm stacks
}

