\name{planar}
\alias{planar}
\title{
	Rcpp module: planar
}
\description{
	Exposes two C++ functions that calculate the reflectivity of a multilayer stack.
}
\details{
	The module contains the following items: 
	
	functions: \describe{
        \item{multilayer}{Calculates the reflection and transmission coefficients of a multilayer stack.}
        \item{recursive_fresnel}{Calculates the reflection and transmission coefficients of a multilayer stack.}
		} 
	
	
}
\keyword{datasets}
