\name{radiationGrob}
\alias{grid.radiation}
\alias{radiationGrob}
\title{radiationGrob}
\usage{
  radiationGrob(theta, r, start = pi, log = TRUE)
}
\arguments{
  \item{theta}{angle}

  \item{r}{radius}

  \item{start}{starting angle}

  \item{log}{logical, log-transformation of radius}
}
\value{
  grob
}
\description{
  radiationGrob
}
\details{
  radiationGrob
}
\author{
  baptiste Auguie
}

