\name{raman.shift}
\alias{raman.shift}
\title{raman.shift}
\usage{raman.shift(lambda=500, shift=1000, stokes=FALSE)}
\description{Raman shift}
\details{calculate the raman shift}
\value{wavelength of the Raman peak in nm}
\author{Baptiste Auguie}
\arguments{\item{lambda}{wavelength (nm)}
\item{shift}{Raman shift (cm-1)}
\item{stokes}{logical Stokes or Anti-Stokes}}
\examples{raman.shift(lambda=200)}
