% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{raman_shift}
\alias{raman_shift}
\title{sort_factor}
\usage{
raman_shift(laser = c(514, 632.8), shift = c(520, 610))
}
\arguments{
\item{laser}{vector of laser wavelengths in nm}

\item{shift}{vector of Raman shifts in cm-1}
}
\value{
matrix of shifted wavelengths (all combinations)
}
\description{
raman_shift
}
\details{
converts Raman shift to wavelength
}
\author{
Baptiste Auguie
}
