% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.r
\name{tamm_stack_ir}
\alias{tamm_stack_ir}
\title{tamm_stack_ir}
\usage{
tamm_stack_ir(lambda0 = 950, n1 = 3, n2 = 3.7, d1 = lambda0/4/n1,
  d2 = lambda0/4/n2, N = 2 * pairs, pairs = 4, dx1 = 0, dx2 = 0,
  dm = 50, metal = "epsAu", position = "after", incidence = "left",
  nleft = n2, nright = 1, ...)
}
\arguments{
\item{lambda0}{central wavelength of the stopband}

\item{n1}{real refractive index for odd layers}

\item{n2}{real refractive index for even layers}

\item{d1}{odd layer thickness in nm}

\item{d2}{even layer thickness in nm}

\item{N}{number of layers, overwrites pairs}

\item{pairs}{number of pairs}

\item{dx1}{variation of last odd layer thickness in nm}

\item{dx2}{variation of last even layer thickness in nm}

\item{dm}{thickness of metal layer}

\item{metal}{character name of dielectric function}

\item{position}{metal position relative to DBR}

\item{incidence}{direction of incidence}

\item{nleft}{refractive index of entering medium}

\item{nright}{refractive index of outer medium}

\item{...}{ignored}
}
\value{
list of class 'stack'
}
\description{
DBR-metal stack structure
}
\details{
periodic structure of dielectric layers against metal film
}
\seealso{
Other stack user_level: \code{\link{dbr_stack}},
  \code{\link{embed_stack}}, \code{\link{layer_stack}},
  \code{\link{tamm_stack_porous}}, \code{\link{tamm_stack}}
}
\author{
baptiste Auguie
}
\concept{stack user_level}
