\name{transmission}
\alias{transmission}
\title{transmission}
\usage{transmission(nPrism, extAngle_r, polar="p")}
\description{transmission loss through a prism}
\details{transmission loss through a prism}
\value{transmission}
\author{baptiste Auguie}
\arguments{\item{nPrism}{prism refractive index}
\item{extAngle_r}{external incident angle in radians}
\item{polar}{polarisation}}
