% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dipole}
\alias{dipole}
\title{dipole}
\usage{
dipole(d = 1, wavelength, epsilon = list(incident = 1^2), thickness = c(0,
  0), qcut = NULL, rel.err = 0.001, Nquadrature1 = 1000,
  Nquadrature2 = 10000, Nquadrature3 = 10000, GL = FALSE,
  show.messages = TRUE)
}
\arguments{
  \item{d}{distance in nm}

  \item{wavelength}{wavelength in nm}

  \item{epsilon}{list of dielectric functions}

  \item{thickness}{list of layer thicknesses}

  \item{qcut}{transition between regions 2 and 3}

  \item{rel.err}{relative error}

  \item{Nquadrature1}{maximum number of quadrature points
  in radiative region}

  \item{Nquadrature2}{maximum number of quadrature points
  in SPPs region}

  \item{Nquadrature3}{maximum number of quadrature points
  in dipole image region}

  \item{GL}{logical: use Gauss Legendre quadrature, or
  cubature::adaptIntegrate}

  \item{show.messages}{logical, display integration info}
}
\description{
Dipole decay rates near a multilayer interface
}
\details{
dipole decay rates near a multilayer interface
}
\author{
baptiste Auguie
}

