% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{epsilon_dispersion}
\alias{epsilon_dispersion}
\title{epsilon_dispersion}
\usage{
epsilon_dispersion(epsilon, wavelength = seq(400, 1000),
  envir = parent.frame())
}
\arguments{
\item{epsilon}{list of real or complex values}

\item{wavelength}{numeric vector}

\item{envir}{environment to look for functions}
}
\value{
list
}
\description{
epsilon_dispersion
}
\details{
apply a function to a range of wavelength and return dielectric function
}
\author{
baptiste Auguie
}

