% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipole.r
\name{integrand_rad}
\alias{integrand_rad}
\title{integrand_rad}
\usage{
integrand_rad(
  d = 10,
  angle,
  wavelength,
  epsilon = list(incident = 1.5^2, 1^2),
  thickness = c(0, 0),
  GL = FALSE
)
}
\arguments{
\item{d}{distance in nm}

\item{angle}{angle in radians}

\item{wavelength}{wavelength in nm}

\item{epsilon}{list of dielectric functions}

\item{thickness}{list of layer thicknesses}

\item{GL}{logical: result formatted for use with Gauss Legendre quadrature}
}
\description{
Dipole decay rates near a multilayer interface
}
\details{
Integrand of the radiative dipole decay rates near a multilayer interface.
}
\seealso{
Other integrands dipole: 
\code{\link{integrand_mtot}()},
\code{\link{integrand_nr1}()},
\code{\link{integrand_nr2}()},
\code{\link{integrand_nr3}()}
}
\author{
baptiste Auguie
}
\concept{integrands dipole}
