\name{invert_stack}
\alias{invert_stack}
\title{invert_stack}
\usage{
  invert_stack(p)
}
\arguments{
  \item{p}{list}
}
\value{
  list
}
\description{
  invert the description of a multilayer to simulate the
  opposite direction of incidence
}
\details{
  inverts list of epsilon and thickness of layers
}
\author{
  Baptiste Auguie
}
\seealso{
  Other helping_functions: \code{\link{field_profile}}
}

