% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{modify_levels}
\alias{modify_levels}
\title{modify_levels}
\usage{
modify_levels(f, modify = list())
}
\arguments{
\item{f}{factor}

\item{modify}{named list}
}
\value{
factor
}
\description{
relabel factors
}
\seealso{
Other helping_functions: 
\code{\link{classify}()},
\code{\link{internal_field}()},
\code{\link{invert_stack}()},
\code{\link{lfief}()}
}
\author{
Baptiste Auguie
}
\concept{helping_functions}
