\name{modify_levels}
\alias{modify_levels}
\title{modify_levels}
\usage{
modify_levels(f, modify = list())
}
\arguments{
  \item{f}{factor}

  \item{modify}{named list}
}
\value{
factor
}
\description{
relabel factors
}
\author{
Baptiste Auguie
}
\seealso{
Other helping_functions: \code{\link{classify}};
\code{\link{field_profile}}, \code{\link{lfief}};
\code{\link{internal_field}}; \code{\link{invert_stack}}
}

