% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{palette_dbr}
\alias{palette_dbr}
\alias{palette_tamm}
\alias{palette_tamm_porous}
\title{Colour palettes for multilayer stacks}
\format{Colour palette (vectors of colours)}
\source{
See RColorBrewer package
}
\description{
Custom palettes for DBR stacks

Custom palette for Tamm stacks

Alternative palette for Tamm stacks
}

