\name{planar-package}
\title{planar-package}
\description{Multilayer optics}
\details{\tabular{ll}{
Package: \tab planar\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-31\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab false\cr
}

R/c++ implementation of the dipole emission near a planar multilayer stack}
\docType{package}
\author{baptiste Auguie \email{baptiste.auguie@gmail.com}}
\references{
Etchegoin, P. Le Ru, E., Principles of Surface-Enhanced Raman Spectroscopy, Elsevier, Amsterdam (2009).
L. Novotny, E. Hecht, Principles of Nano-optics Cambridge University Press, 2006 
H. Raether. Surface Plasmons on Smooth and Rough Surfaces and on Gratings. Springer, 1988.
}
\keyword{packagelibrary}
