\name{radiationGrob}
\title{radiationGrob}
\usage{radiationGrob(theta, r, start=pi, log=TRUE)}
\description{radiationGrob}
\details{radiationGrob}
\alias{grid.radiation}
\alias{radiationGrob}
\value{grob}
\author{baptiste Auguie}
\arguments{\item{theta}{angle}
\item{r}{radius}
\item{start}{starting angle}
\item{log}{logical, log-transformation of radius}}
