% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.r
\name{rev.stack}
\alias{rev.stack}
\title{rev.stack}
\usage{
\method{rev}{stack}(x)
}
\arguments{
\item{x}{stack}
}
\value{
stack
}
\description{
invert the description of a multilayer to simulate the opposite direction of incidence
}
\details{
inverts list of epsilon and thickness of layers
}
\seealso{
Other helping_functions user_level stack: 
\code{\link{simulate_ff}()},
\code{\link{simulate_nf}()}
}
\author{
Baptiste Auguie
}
\concept{helping_functions user_level stack}
