% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_nf}
\alias{simulate_nf}
\title{simulate_nf}
\usage{
simulate_nf(..., s = NULL, fun = tamm_stack, wavelength = 630,
  angle = 0, polarisation = c("p", "s"), dmax = 0, res = 10000,
  field = FALSE)
}
\arguments{
\item{...}{further arguments passed to fun}

\item{s}{stack (optional)}

\item{fun}{function returning a stack}

\item{wavelength}{numeric vector}

\item{angle}{incident angle in radians}

\item{polarisation}{p or s}

\item{dmax}{maximum distance from stack boundary}

\item{res}{number of points}

\item{field}{logical, return the real electric field}
}
\value{
data.frame
}
\description{
simultate the internal field of a multilayer stack
}
\details{
wrapper around multilayer_field for a stack structure
}
\seealso{
Other helping_functions user_level stack: \code{\link{rev.stack}},
  \code{\link{simulate_ff}}
}
\author{
Baptiste Auguie
}
\concept{helping_functions user_level stack}
