\name{theme_minimal}
\alias{theme_minimal}
\title{theme_minimal}
\usage{
  theme_minimal(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{font size}

  \item{base_family}{font family}
}
\value{
  theme (list)
}
\description{
  minimal theme
}
\details{
  bare-bones theme for ggplot2
}
\author{
  baptiste Auguie
}

