\name{transmission}
\alias{transmission}
\title{transmission}
\usage{
  transmission(nPrism, extAngle_r, polar = "p")
}
\arguments{
  \item{nPrism}{prism refractive index}

  \item{extAngle_r}{external incident angle in radians}

  \item{polar}{polarisation}
}
\value{
  transmission
}
\description{
  transmission loss through a prism
}
\details{
  transmission loss through a prism
}
\author{
  baptiste Auguie
}

