% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.r
\name{transmission}
\alias{transmission}
\title{transmission}
\usage{
transmission(n, external, polarisation = "p")
}
\arguments{
\item{n}{prism refractive index}

\item{external}{external incident angle in radians}

\item{polarisation}{polarisation}
}
\value{
transmission
}
\description{
transmission loss through a prism
}
\details{
transmission loss through a prism
}
\author{
baptiste Auguie
}
